Removing comments and unused imports

+++ apis/room/controllers/delete.js
@@ -12,12 +12,14 @@ const deleteRoomByID = (req, res) => {
 };
 
 const deleteRoomsByUserID = async (req, res) => {
+  if (!(req.isAuthenticated() && req.user._id == req.body.user._id)) {
+    return res.status(403).send();
+  }
   const result = await Room.deleteMany({ ownerID: req.user._id });
   res.status(200).send(result);
 };
 
 const deleteUserFromRoom = async (req, res) => {  
-
   if (!(req.isAuthenticated() && req.user._id == req.body.user._id)) {
     return res.status(403).send();
   }
